Feat: Base for new billing settings
