+++ packages/frontend-2/components/settings/workspaces/billing/AddOns.vue
@@ -1,0 +1,21 @@
+<!-- TODO: Implement final values, links and copy -->
+<template>
+  <div class="border border-outline-3 rounded-lg divide-x flex flex-col md:flex-row">
+    <div class="p-6 w-1/2">
+      <h3 class="text-body font-medium">Unlimited projects and models</h3>
+      <p class="text-body-2xs text-foreground-2 mt-1">plus X per seat / month</p>
+      <p class="max-w-60 text-foreground-2 text-body-2xs mt-6">
+        Add more projects and models to your workspace.
+      </p>
+      <FormButton class="mt-6" color="outline" size="sm">Contact us</FormButton>
+    </div>
+    <div class="p-6 w-1/2">
+      <h3 class="text-body font-medium">Extra data regions</h3>
+      <p class="text-body-2xs text-foreground-2 mt-1">Talk to us!</p>
+      <p class="max-w-60 text-foreground-2 text-body-2xs mt-6">
+        Add more available data regions. Learn more about Speckle Data Regions.
+      </p>
+      <FormButton class="mt-6" color="outline" size="sm">Contact us</FormButton>
+    </div>
+  </div>
+</template>
