+++ packages/frontend-2/components/settings/workspaces/billing/usage/index.vue
@@ -1,0 +1,25 @@
+<!-- TODO: Update these with real limits once available -->
+<template>
+  <div class="flex flex-col gap-y-4">
+    <SettingsWorkspacesBillingUsageCard
+      button-text="Manage seats"
+      :current-value="7"
+      :max-value="8"
+      type="seat"
+    />
+    <SettingsWorkspacesBillingUsageCard
+      button-text="Manage projects"
+      :current-value="7"
+      :max-value="50"
+      type="project"
+    />
+    <SettingsWorkspacesBillingUsageCard
+      button-text="Manage versions"
+      :current-value="50"
+      :max-value="50"
+      type="version"
+    />
+  </div>
+</template>
+
+<script lang="ts" setup></script>
