+++ packages/frontend-2/lib/workspaces/composables/plan.ts
@@ -1,0 +1,115 @@
+import { graphql } from '~~/lib/common/generated/gql'
+import { workspacePlanQuery } from '~~/lib/workspaces/graphql/queries'
+import { useQuery } from '@vue/apollo-composable'
+import {
+  isNewWorkspacePlan,
+  PaidWorkspacePlansNew,
+  TrialEnabledPaidWorkspacePlans
+} from '@speckle/shared'
+import {
+  WorkspacePlanStatuses,
+  BillingInterval
+} from '~/lib/common/generated/gql/graphql'
+
+graphql(`
+  fragment WorkspacesPlan_Workspace on Workspace {
+    id
+    plan {
+      status
+      createdAt
+      name
+    }
+    subscription {
+      billingInterval
+    }
+  }
+`)
+
+export const useWorkspacePlan = (slug: string) => {
+  const isBillingIntegrationEnabled = useIsBillingIntegrationEnabled()
+
+  const { result } = useQuery(
+    workspacePlanQuery,
+    () => ({
+      slug
+    }),
+    () => ({
+      enabled: isBillingIntegrationEnabled
+    })
+  )
+
+  const subscription = computed(() => result.value?.workspaceBySlug?.subscription)
+  const plan = computed(() => result.value?.workspaceBySlug?.plan)
+
+  const isNewPlan = computed(() =>
+    isNewWorkspacePlan(result.value?.workspaceBySlug?.plan?.name)
+  )
+
+  const statusIsExpired = computed(
+    () => plan.value?.status === WorkspacePlanStatuses.Expired
+  )
+
+  const statusIsCanceled = computed(
+    () => plan.value?.status === WorkspacePlanStatuses.Canceled
+  )
+
+  const statusIsCancelationScheduled = computed(
+    () => plan.value?.status === WorkspacePlanStatuses.CancelationScheduled
+  )
+
+  const isPurchasablePlan = computed(() =>
+    Object.values(PaidWorkspacePlansNew).includes(
+      plan.value?.name as PaidWorkspacePlansNew
+    )
+  )
+
+  const isActivePlan = computed(
+    () =>
+      plan.value?.status === WorkspacePlanStatuses.Valid ||
+      plan.value?.status === WorkspacePlanStatuses.PaymentFailed ||
+      plan.value?.status === WorkspacePlanStatuses.CancelationScheduled
+  )
+
+  const isFreePlan = computed(
+    () => plan.value?.name === TrialEnabledPaidWorkspacePlans.Starter
+  )
+
+  const billingInterval = computed(() => subscription.value?.billingInterval)
+
+  const intervalIsYearly = computed(
+    () => billingInterval.value === BillingInterval.Yearly
+  )
+
+  // TODO: Replace with value from API call, this a placeholder value
+  const seatPrice = 15
+
+  const totalCost = computed(() => {
+    return isPurchasablePlan.value
+      ? intervalIsYearly.value
+        ? seatPrice * 12
+        : seatPrice
+      : 0
+  })
+
+  // TODO: Replace with value from BE once ready
+  const totalCostFormatted = computed(() => {
+    return isPurchasablePlan.value
+      ? `£${totalCost.value}`
+      : isFreePlan.value
+      ? 'Free'
+      : 'Not applicable'
+  })
+
+  return {
+    plan,
+    isNewPlan,
+    statusIsExpired,
+    statusIsCanceled,
+    isPurchasablePlan,
+    isActivePlan,
+    billingInterval,
+    intervalIsYearly,
+    totalCostFormatted,
+    statusIsCancelationScheduled
+  }
+}
