+++ packages/frontend-2/lib/workspaces/graphql/queries.ts
@@ -125,7 +125,7 @@ export const workspaceWizardRegionQuery = graphql(`
 `)
 
 export const discoverableWorkspacesQuery = graphql(`
-  query DiscoverableWorkspaces_ActiveUser {
+  query DiscoverableWorkspaces {
     activeUser {
       id
       ...DiscoverableList_Discoverable
@@ -134,10 +134,18 @@ export const discoverableWorkspacesQuery = graphql(`
 `)
 
 export const discoverableWorkspacesRequestsQuery = graphql(`
-  query DiscoverableWorkspacesRequests_ActiveUser {
+  query DiscoverableWorkspacesRequests {
     activeUser {
       id
       ...DiscoverableList_Requests
+    }
+  }
+`)
+
+export const workspacePlanQuery = graphql(`
+  query WorkspacePlan($slug: String!) {
+    workspaceBySlug(slug: $slug) {
+      ...WorkspacesPlan_Workspace
     }
   }
 `)
